PCHR-1690: Fixes job role UI updates

+++ com.civicrm.hrjobroles/js/src/job-roles/controllers/hr-job-roles-controller.js
@@ -8,10 +8,10 @@ define([
   'use strict';
 
   controllers.controller('HRJobRolesController', [
-    '$scope', '$log', '$routeParams', '$route', '$timeout', '$filter', '$q',
+    '$scope', '$log', '$routeParams', '$route', '$uibModal', '$rootElement', '$timeout', '$filter', '$q','settings',
     'HR_settings', 'HRJobRolesService', 'DateValidation', 'HRJobRolesServiceFilters',
     'DOMEventTrigger',
-    function ($scope, $log, $routeParams, $route, $timeout, $filter, $q, HR_settings, HRJobRolesService, DateValidation, HRJobRolesServiceFilters, DOMEventTrigger) {
+    function ($scope, $log, $routeParams, $route, $modal, $rootElement, $timeout, $filter, $q, settings, HR_settings, HRJobRolesService, DateValidation, HRJobRolesServiceFilters, DOMEventTrigger) {
       $log.debug('Controller: HRJobRolesController');
 
       $scope.format = HR_settings.DATE_FORMAT;
@@ -22,6 +22,11 @@ define([
 
       $scope.present_job_roles = [];
       $scope.past_job_roles = [];
+      $scope.templateLoaded = false;
+
+      $scope.loaded = function(){
+        $scope.templateLoaded = true;
+      }
 
       $scope.dpOpen = function ($event) {
         $event.preventDefault();
@@ -223,7 +228,6 @@ define([
 
       // Set the data from the webservice call
       $scope.initData = function (role_id, form_id, data) {
-
         // Check if we have the array already
         if (typeof $scope.edit_data[role_id] === "undefined") {
           $scope.edit_data[role_id] = {};
@@ -492,12 +496,33 @@ define([
       $scope.removeRole = function (jobRole) {
         $log.debug('Remove Role');
 
-        // Delete job role
-        deleteJobRole(jobRole.id).then(function () {
-          updateHeaderInfo(jobRole);
-
-          return getJobRolesList($scope.$parent.contactId);
+        var modalInstance = $modal.open({
+            appendTo: $rootElement.find('div').eq(0),
+            template: '',
+            templateUrl: settings.pathApp+'views/modalDialog.html?v='+(new Date()).getTime(),
+            size: 'sm',
+            controller: 'ModalDialogCtrl',
+            resolve: {
+                content: function(){
+                    return {
+                        copyCancel: 'No',
+                        title: 'Alert',
+                        msg: 'Are you sure you want to Delete Job Role?'
+                    };
+                }
+            }
         });
+
+        // Delete job role
+        modalInstance.result.then(function (confirm) {
+          if (confirm) {
+            deleteJobRole(jobRole.id).then(function () {
+              updateHeaderInfo(jobRole);
+
+              return getJobRolesList($scope.$parent.contactId);
+            });
+          }
+        })
       };
 
       /**
@@ -683,6 +708,7 @@ define([
       // Get the option groups and option values
       getOptionValues();
 
+      $scope.$broadcast('ct-spinner-show');
       // Get job roles based on the passed Contact ID
       getJobRolesList($scope.$parent.contactId);
